refactor(BUX-612): interface.ts to types.ts
